+++ packages/utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Added `Flags.string({ optionalValue })`: a bare `--flag` takes the stated value ([#12630](https://github.com/can1357/oh-my-pi/pull/12630) by [@ParadaCarleton](https://github.com/ParadaCarleton)).
+
 ## [18.3.1] - 2026-09-25
 
 ### Fixed